The most significant changes involve the modification of hyperlink generation in `Index.razor` and the addition of a new file `StringExt.cs` in the `PNetRunner.Ext` namespace.
In `Index.razor`, the hyperlink was previously created by directly replacing "-S " with an empty string in the `StartInfo.Arguments` property of `proc.Item2`. The updated code now uses a new extension method `AsHref` to generate the hyperlink. This method takes a boolean parameter to specify whether "http://" should be added to the string, and a list of strings to be replaced in the input string.

The new file `StringExt.cs` has been added to the `PNetRunner.Ext` namespace. This file contains a static class `StringExt` that provides extension methods for string manipulation. The `AsHref` method in this class converts a string to a hyperlink reference (href) format. It takes an input string, a boolean to specify whether "http://" should be added to the string, and a list of strings to be replaced in the input string. The method then replaces each string in the list with an empty string in the input string, and if the boolean is true, it adds "http://" to the start of the string. The method then returns the modified string.

Changes:
1. Modification of hyperlink generation in `Index.razor` to use the new `AsHref` extension method.
2. Addition of a new file `StringExt.cs` in the `PNetRunner.Ext` namespace, which contains the `StringExt` static class and the `AsHref` extension method.